Remove app start test from react native tracing
